(TEST) bbram_generic test
test board: it8xxx2, and it82xx2_evb(it82202bw)

mcp7940 test is failed.

+++ include/zephyr/drivers/emul.h
@@ -68,7 +68,7 @@ typedef int (*emul_init_t)(const struct emul *emul, const struct device *parent)
  */
 struct no_bus_emul {
 	void *api;
-	uint16_t addr;
+	int addr;
 };
 
 /** An emulator instance - represents the *target* emulated device/peripheral that is
